requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-extensions==3.1.5
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
djangorestframework-stubs==1.6.0
google-play-scraper==1.2.2
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy==0.950
mypy-extensions==0.4.3
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
tomli==2.0.1
types-Markdown==3.3.28
types-pytz==2021.3.8
types-PyYAML==6.0.8
types-requests==2.27.30
types-urllib3==1.26.15
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
beautifulsoup4==4.11.1
nltk==3.7
app-store-scraper==0.3.5